+++ upload/admin/controller/localisation/weight_class.php
@@ -155,7 +155,7 @@ class ControllerLocalisationWeightClass extends Controller {
 			'href' => $this->url->link('localisation/weight_class', 'token=' . $this->session->data['token'] . $url, 'SSL')
 		);
 
-		$data['insert'] = $this->url->link('localisation/weight_class/add', 'token=' . $this->session->data['token'] . $url, 'SSL');
+		$data['add'] = $this->url->link('localisation/weight_class/add', 'token=' . $this->session->data['token'] . $url, 'SSL');
 		$data['delete'] = $this->url->link('localisation/weight_class/delete', 'token=' . $this->session->data['token'] . $url, 'SSL');
 
 		$data['weight_classes'] = array();
@@ -192,7 +192,7 @@ class ControllerLocalisationWeightClass extends Controller {
 		$data['column_value'] = $this->language->get('column_value');
 		$data['column_action'] = $this->language->get('column_action');
 
-		$data['button_insert'] = $this->language->get('button_insert');
+		$data['button_add'] = $this->language->get('button_add');
 		$data['button_edit'] = $this->language->get('button_edit');
 		$data['button_delete'] = $this->language->get('button_delete');