+++ upload/admin/controller/openbay/amazon.php
@@ -6,15 +6,15 @@ class ControllerOpenbayAmazon extends Controller {
 		$this->load->model('extension/extension');
 		$this->load->model('extension/event');
 
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'openbay/amazon_listing');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'openbay/amazon_listing');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'openbay/amazon_product');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'openbay/amazon_product');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'access', 'openbay/amazon_listing');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'modify', 'openbay/amazon_listing');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'access', 'openbay/amazon_product');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'modify', 'openbay/amazon_product');
 
-		$this->model_openbay_amazon->install();
+		$this->model_openbay_amazon->install(); 
 	}
 
-	public function uninstall() {
+	public function uninstall() {  
 		$this->load->model('openbay/amazon');
 		$this->load->model('setting/setting');
 		$this->load->model('extension/extension');