+++ upload/admin/controller/openbay/amazonus.php
@@ -6,10 +6,10 @@ class ControllerOpenbayAmazonus extends Controller {
 		$this->load->model('extension/extension');
 		$this->load->model('extension/event');
 
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'openbay/amazonus_listing');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'openbay/amazonus_listing');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'openbay/amazonus_product');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'openbay/amazonus_product');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'access', 'openbay/amazonus_listing');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'modify', 'openbay/amazonus_listing');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'access', 'openbay/amazonus_product');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'modify', 'openbay/amazonus_product');
 
 		$this->model_openbay_amazonus->install();
 	}