+++ upload/admin/controller/openbay/ebay.php
@@ -7,10 +7,10 @@ class ControllerOpenbayEbay extends Controller {
 		$this->load->model('extension/extension');
 		$this->load->model('extension/event');
 
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'openbay/ebay_profile');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'openbay/ebay_profile');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'openbay/ebay_template');
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'openbay/ebay_template');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'access', 'openbay/ebay_profile');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'modify', 'openbay/ebay_profile');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'access', 'openbay/ebay_template');
+		$this->model_user_user_group->addPermission($this->user->getGroupId(), 'modify', 'openbay/ebay_template');
 
 		$this->model_openbay_ebay->install();
 	}
@@ -125,13 +125,6 @@ class ControllerOpenbayEbay extends Controller {
 			$data['ebay_def_currency'] = $this->config->get('ebay_def_currency');
 		}
 		$data['currency_list'] = $this->model_localisation_currency->getCurrencies();
-
-		if (isset($this->request->post['ebay_def_customer_grp'])) {
-			$data['ebay_def_customer_grp'] = $this->request->post['ebay_def_customer_grp'];
-		} else {
-			$data['ebay_def_customer_grp'] = $this->config->get('ebay_def_customer_grp');
-		}
-		$data['customer_grp_list'] = $this->model_sale_customer_group->getCustomerGroups();
 
 		$data['token'] = $this->session->data['token'];
 
@@ -1291,7 +1284,7 @@ class ControllerOpenbayEbay extends Controller {
 				$product_info['product_images'] = array();
 
 				if (!empty($product_info['image'])) {
-					$img_info = getimagesize(HTTPS_CATALOG . 'image/' . $product_info['image']);
+					$img_info = getimagesize(DIR_IMAGE . $product_info['image']);
 
 					$product_info['product_images'][] = array(
 						'image' => $product_info['image'],
@@ -1304,7 +1297,7 @@ class ControllerOpenbayEbay extends Controller {
 
 				foreach ($product_images as $product_image) {
 					if ($product_image['image'] && file_exists(DIR_IMAGE . $product_image['image'])) {
-						$img_info = getimagesize(HTTPS_CATALOG . 'image/' . $product_image['image']);
+						$img_info = getimagesize(DIR_IMAGE . $product_image['image']);
 
 						$product_info['product_images'][] = array(
 							'image' => $product_image['image'],