+++ upload/admin/controller/setting/store.php
@@ -99,7 +99,7 @@ class ControllerSettingStore extends Controller {
 			'href' => $this->url->link('setting/store', 'token=' . $this->session->data['token'], 'SSL')
 		);
 
-		$data['insert'] = $this->url->link('setting/store/add', 'token=' . $this->session->data['token'], 'SSL');
+		$data['add'] = $this->url->link('setting/store/add', 'token=' . $this->session->data['token'], 'SSL');
 		$data['delete'] = $this->url->link('setting/store/delete', 'token=' . $this->session->data['token'], 'SSL');
 
 		$data['stores'] = array();
@@ -134,7 +134,7 @@ class ControllerSettingStore extends Controller {
 		$data['column_url'] = $this->language->get('column_url');
 		$data['column_action'] = $this->language->get('column_action');
 
-		$data['button_insert'] = $this->language->get('button_insert');
+		$data['button_add'] = $this->language->get('button_add');
 		$data['button_edit'] = $this->language->get('button_edit');
 		$data['button_delete'] = $this->language->get('button_delete');
 
@@ -405,12 +405,19 @@ class ControllerSettingStore extends Controller {
 			'text' => $this->language->get('heading_title'),
 			'href' => $this->url->link('setting/store', 'token=' . $this->session->data['token'], 'SSL')
 		);
-
-		$data['breadcrumbs'][] = array(
-			'text' => $this->language->get('text_settings'),
-			'href' => $this->url->link('setting/store/edit', 'token=' . $this->session->data['token'] . '&store_id=' . $this->request->get['store_id'], 'SSL')
-		);
-
+		
+		if (!isset($this->request->get['store_id'])) {
+			$data['breadcrumbs'][] = array(
+				'text' => $this->language->get('text_settings'),
+				'href' => $this->url->link('setting/store/add', 'token=' . $this->session->data['token'], 'SSL')
+			);
+		} else {
+			$data['breadcrumbs'][] = array(
+				'text' => $this->language->get('text_settings'),
+				'href' => $this->url->link('setting/store/edit', 'token=' . $this->session->data['token'] . '&store_id=' . $this->request->get['store_id'], 'SSL')
+			);			
+		}
+		
 		if (isset($this->session->data['success'])) {
 			$data['success'] = $this->session->data['success'];
 
@@ -508,11 +515,11 @@ class ControllerSettingStore extends Controller {
 		}
 
 		if (isset($this->request->post['config_image'])) {
-			$data['image'] = $this->request->post['config_image'];
+			$data['config_image'] = $this->request->post['config_image'];
 		} elseif (isset($store_info['config_image'])) {
-			$data['image'] = $store_info['config_image'];
+			$data['config_image'] = $store_info['config_image'];
 		} else {
-			$data['image'] = '';
+			$data['config_image'] = '';
 		}
 
 		$this->load->model('tool/image');
@@ -527,20 +534,20 @@ class ControllerSettingStore extends Controller {
 
 		$data['placeholder'] = $this->model_tool_image->resize('no_image.png', 100, 100);
 
-		if (isset($this->request->post['open'])) {
-			$data['open'] = $this->request->post['open'];
+		if (isset($this->request->post['config_open'])) {
+			$data['config_open'] = $this->request->post['config_open'];
 		} elseif (isset($store_info['config_open'])) {
-			$data['open'] = $store_info['open'];
+			$data['config_open'] = $store_info['config_open'];
 		} else {
-			$data['open'] = '';
+			$data['config_open'] = '';
 		}
 
-		if (isset($this->request->post['comment'])) {
-			$data['comment'] = $this->request->post['comment'];
+		if (isset($this->request->post['config_comment'])) {
+			$data['config_comment'] = $this->request->post['config_comment'];
 		} elseif (isset($store_info['config_comment'])) {
-			$data['comment'] = $store_info['comment'];
+			$data['config_comment'] = $store_info['config_comment'];
 		} else {
-			$data['comment'] = '';
+			$data['config_comment'] = '';
 		}
 
 		$this->load->model('localisation/location');