+++ upload/catalog/controller/module/category.php
@@ -1,6 +1,6 @@
 <?php
 class ControllerModuleCategory extends Controller {
-	public function index($setting) {
+	public function index() {
 		$this->load->language('module/category');
 
 		$data['heading_title'] = $this->language->get('heading_title');
@@ -40,7 +40,11 @@ class ControllerModuleCategory extends Controller {
 				foreach($children as $child) {
 					$filter_data = array('filter_category_id' => $child['category_id'], 'filter_sub_category' => true);
 
-					$children_data[] = array('category_id' => $child['category_id'], 'name' => $child['name'] . ($this->config->get('config_product_count') ? ' (' . $this->model_catalog_product->getTotalProducts($filter_data) . ')' : ''), 'href' => $this->url->link('product/category', 'path=' . $category['category_id'] . '_' . $child['category_id']));
+					$children_data[] = array(
+						'category_id' => $child['category_id'], 
+						'name' => $child['name'] . ($this->config->get('config_product_count') ? ' (' . $this->model_catalog_product->getTotalProducts($filter_data) . ')' : ''), 
+						'href' => $this->url->link('product/category', 'path=' . $category['category_id'] . '_' . $child['category_id'])
+					);
 				}
 			}