+++ upload/catalog/model/setting/setting.php
@@ -1,9 +1,9 @@
 <?php
 class ModelSettingSetting extends Model {
-	public function getSetting($group, $store_id = 0) {
+	public function getSetting($code, $store_id = 0) {
 		$data = array();
 
-		$query = $this->db->query("SELECT * FROM " . DB_PREFIX . "setting WHERE store_id = '" . (int)$store_id . "' AND `group` = '" . $this->db->escape($group) . "'");
+		$query = $this->db->query("SELECT * FROM " . DB_PREFIX . "setting WHERE store_id = '" . (int)$store_id . "' AND `code` = '" . $this->db->escape($code) . "'");
 
 		foreach ($query->rows as $result) {
 			if (!$result['serialized']) {